fix: 修复所有16类编译错误，适配Unity 6 API
- GameEvents: 移除event关键字，改为public static Action委托
- TextAlignmentOptions: MiddleCenter→Center, MiddleLeft→MidlineLeft, MiddleRight→MidlineRight
- StarMap: Tuple<TMP_Text>→Tuple<TextMeshProUGUI>
- ObstacleCollisionSystem: 添加using WeiJinRoad.World, uint→int转换
- FogSystem/SnowSystem: 移除billboardMode行
- SceneBootstrap: light.shadowDistance→QualitySettings.shadowDistance
- VehicleController: HeadlightsMode→HeadlightsModeType, SpendFuel→SpendResources
- SceneryElements: CreateConeMesh改为public, AddCylinder添加emissiveColor参数
- EnvironmentLighting: 移除LightType.Ambient引用
- ResourceNodes/TownScene: AchievementSystem静态调用改为FindObjectOfType实例调用
- TerrainGenerator: 添加using WeiJinRoad.Vehicle, Vector2.z→Vector2.y
- RepairMenu: contentArea→contentRect
- DevToolsUI: 移除UnityStats引用

+++ Assets/Scripts/Collision/ObstacleCollisionSystem.cs
@@ -18,6 +18,7 @@ using UnityEngine;
 using WeiJinRoad.Core;
 using WeiJinRoad.Data;
 using WeiJinRoad.Vehicle;
+using WeiJinRoad.World;
 
 namespace WeiJinRoad.Collision
 {
@@ -620,7 +621,7 @@ namespace WeiJinRoad.Collision
         /// <returns>障碍定义列表</returns>
         private static List<ObstacleDef> GenerateObstacles()
         {
-            var rng = new Mulberry32(GameData.ObstacleSeed);
+            var rng = new Mulberry32((int)GameData.ObstacleSeed);
             var list = new List<ObstacleDef>();
             int index = 0;
